moves social package to /login
ref #14679
